User Testing
Tested .apk on a Masstel_Tab10 device. Running Android 8. It said that i needed goolge play services for AR. But wouldn't allow me to update it. I Could see the start screen. The app would open but when i clicked on any of the shoes, i got an error message saying. "Incompatible with this device."